+++ src/features/Score.tsx
